+++ test/locales/translations.test.js
@@ -242,7 +242,8 @@ const ignoredTransKeyInYourLoopsFiles = [
   'Pump cartridge expiration date',
   'Pump version',
   'Serial Number',
-  'Setting'
+  'Setting',
+  'Time In Range'
 ]
 
 /**